pyproject.toml
[tool.poetry]
name = "homework-24-1"
version = "0.1.0"
description = ""
authors = ["Sprite_Spirit <Sprite.spirit.t.me@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
pillow = "^10.4.0"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
django-phonenumber-field = "^8.0.0"
django-countries = "^7.6.1"
django-phonenumbers = "^1.0.1"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
requests = "^2.32.3"
stripe = "^10.8.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
